[Behat] Extract ProductChannelPricingsTrait from product pages
